requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Bottleneck==1.4.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
clarabel==0.9.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cvxpy==1.6.0
cycler==0.12.1
debugpy==1.8.10
decorator==5.1.1
ecos==2.0.14
empyrical-reloaded==0.5.11
executing==2.1.0
fastjsonschema==2.21.1
fonttools==4.55.3
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
nltp==0.1.0
numpy==2.2.0
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
peewee==3.17.3
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pydeck==0.9.1
pyfolio-reloaded==0.9.8
Pygments==2.18.0
pyparsing==3.2.0
pyportfolioopt==1.5.6
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
qdldl==0.1.7.post4
QuantStats==0.0.64
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.14.1
scs==3.2.7
seaborn==0.13.2
six==1.17.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.41.0
ta-lib @ file:///C:/Users/Darkles/10academy/10Academy-Kifiya-Week-1/ta_lib-0.5.1-cp312-cp312-win_amd64.whl#sha256=a691eda5ce91c0436742c366ad578dd9936e71fd0ee411d56979a3ad696e20c3
tabulate==0.9.0
tenacity==9.0.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
windrose==1.9.2
wordcloud==1.9.4
yfinance==0.2.50
